backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
celery
prometheus-fastapi-instrumentator
boto3
python-jose[cryptography]
passlib[bcrypt]
asyncpg
aioredis
fastapi-cache2
slowapi
